add create and drop table, left some problem to check

+++ src/system/TableManager.cpp
@@ -6,13 +6,24 @@ TableManager::TableManager(string databaseName_,  BufPageManager* bufPageManager
     recordManager = new RecordManager(bufPageManager_, databaseName.c_str());
 }
 
-bool TableManager::checkTableExist(string path) {
+inline bool TableManager::checkTableName(string name) {
+    size_t length = name.length();
+    if(length == 0 || length > TAB_MAX_NAME_LEN) {
+        printf("[Error] invalid table name ! \n");
+        return false;
+    }
+    return true;
+}
+
+inline bool TableManager::checkTableExist(string path) {
     if (!access(path.c_str(), F_OK))
         return true; // already exit
     return false;
 }
 
 int TableManager::creatTable(string name) {
+    if(!checkTableName(name))
+        return -1;
     string path = "database/" + databaseName + '/' + name +".db";
     if(!checkTableExist(path))
         if(recordManager->createFile(name.c_str()) == 0)
@@ -22,6 +33,8 @@ int TableManager::creatTable(string name) {
 }
 
 int TableManager::openTable(string name) {
+    if(!checkTableName(name))
+        return -1;
     string path = "database/" + databaseName + '/' + name +".db";
     FileHandler* f;
     if(checkTableExist(path))
@@ -32,6 +45,8 @@ int TableManager::openTable(string name) {
 }
 
 int TableManager::dropTable(string name) {
+    if(!checkTableName(name))
+        return -1;
     string path = "database/" + databaseName + '/' + name +".db";
     if(checkTableExist(path))
         if(recordManager->removeFile(name.c_str()) == 0)
